requirements.txt
# Core 
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0

# CLI 
click>=8.1.0

# Machine learning
scikit-learn>=1.8.0
xgboost>=3.2.0
joblib>=1.5.3
scipy>=1.17.0
# optional (for hyperparameter optimization)
optuna>=4.7.0

# Testing 
pytest>=9.0.2

# Data sources
MLB-StatsAPI>=1.9.0
